pyproject.toml
[project]
name = "gigi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.14",
    "aiosqlite>=0.21.0",
    "msgspec>=0.19.0",
    "python-dotenv>=1.1.1",
    "starlette>=0.47.1",
    "uvicorn>=0.35.0",
]
